Made change in the register page
