+++ package.json
@@ -13,6 +13,7 @@
     "@tomtom-international/web-sdk-maps": "^6.23.0",
     "@tomtom-international/web-sdk-services": "^6.23.0",
     "axios": "^1.3.3",
+    "bcrypt": "^5.1.0",
     "dotenv": "^16.0.3",
     "framer-motion": "^9.0.2",
     "prop-types": "^15.8.1",
