+++ src/pages/LoginPage.jsx
@@ -1,15 +1,121 @@
-import React from "react";
-import { Box } from "@chakra-ui/react";
-import Login from "../components/Login/Login";
-import { useDisclosure } from "@chakra-ui/react";
+import {
+  Box,
+  FormControl,
+  FormLabel,
+  Input,
+  InputGroup,
+  InputRightElement,
+  Stack,
+  Button,
+  Text,
+  Link,
+  Image,
+} from "@chakra-ui/react";
+import { useState } from "react";
+import { ViewIcon, ViewOffIcon } from "@chakra-ui/icons";
+import yasai from "../assets/yasai.png";
+import { useNavigate } from "react-router-dom";
+import { useToast } from "@chakra-ui/react";
+import axios from "axios";
+
+const API_URL = "http://localhost:8080/auth/login";
+
 const LoginPage = () => {
-  const initialRef = React.useRef(null);
-  const finalRef = React.useRef(null);
-  const login = useDisclosure();
+  const [showPassword, setShowPassword] = useState(false);
+  const [email, setEmail] = useState("");
+  const [password, setPassword] = useState("");
+  const nav = useNavigate();
+  const toast = useToast();
+
+  const handleSubmit = () => {
+    axios
+      .post(API_URL, {
+        email: email,
+        password: password,
+      })
+      .then((response) => {
+        console.log(response);
+        toast({
+          title: "Logged in successfully.",
+          description: "Redirecting to next page...",
+          status: "success",
+          duration: 9000,
+          position: "top",
+          isClosable: true,
+        });
+        nav("/login");
+      })
+      .catch((error) => {
+        console.log(error.messsage);
+        toast({
+          title: "Account has errors.",
+          description: "Check your credentials.",
+          status: "error",
+          duration: 9000,
+          position: "top",
+          isClosable: true,
+        });
+      });
+  };
+
   return (
     <>
       <Box display="flex" justifyContent="center">
-        <Login />
+        <Box transform="translate(0%, 0%)" w={[300, 400, 500, 800]} p="5">
+          <Image src={yasai} alt="yasai" />
+          <Stack spacing={4}>
+            <FormControl id="email" isRequired>
+              <FormLabel>Email</FormLabel>
+              <Input
+                onChange={(e) => setEmail(e.target.value)}
+                value={email}
+                type="email"
+              />
+            </FormControl>
+            <FormControl id="password" isRequired>
+              <FormLabel>Password</FormLabel>
+              <InputGroup>
+                <Input
+                  onChange={(e) => setPassword(e.target.value)}
+                  value={password}
+                  type={showPassword ? "text" : "password"}
+                />
+                <InputRightElement h={"full"}>
+                  <Button
+                    variant={"ghost"}
+                    onClick={() =>
+                      setShowPassword((showPassword) => !showPassword)
+                    }
+                  >
+                    {showPassword ? <ViewIcon /> : <ViewOffIcon />}
+                  </Button>
+                </InputRightElement>
+              </InputGroup>
+            </FormControl>
+            <Stack spacing={10} pt={2}>
+              <Button
+                onClick={handleSubmit}
+                loadingText="Submitting"
+                size="lg"
+                bg={"green.400"}
+                color={"white"}
+                _hover={{
+                  bg: "blue.500",
+                }}
+              >
+                Sign up
+              </Button>
+            </Stack>
+            <Stack pt={6}>
+              <Text align={"center"}>
+                Not Registered yet ?{" "}
+                <Link color={"blue.400"} href="/register">
+                  Register
+                </Link>
+              </Text>
+            </Stack>
+          </Stack>
+        </Box>
       </Box>
     </>
   );
