+++ src/pages/ShopPage/ShopView.jsx
@@ -15,7 +15,6 @@ import {
   DrawerHeader,
   DrawerFooter,
   Button,
-  Input,
   ButtonGroup,
   FormControl,
   FormLabel,
